Fix bead lifecycle error handling
